refactor: move filter components to search/filters
