+++ frontend/src/components/search/filters/Filters.scss
